requirements.txt
flask
uWSGI
mysql-connector-python
